+++ docs/source/conf.py
@@ -27,7 +27,7 @@ author = "Clement Julia"
 # The short X.Y version
 version = ""
 # The full version, including alpha/beta/rc tags
-release = "0.9.1"
+release = "0.11.0"
 
 
 # -- General configuration ---------------------------------------------------