+++ moddb/pages/entity.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
-import logging
 
+import logging
 from typing import TYPE_CHECKING, List, Tuple
 
 from ..boxes import (